Fixing Underflow of ps_dec->u2_num_mbs_left
In multi-thread decode, the decoder would try to decode without
dequeueing a job in case the next slice indicated that it belongs to the
same row as being decoded currently. In single thread case, there was a
check to ensure that the decoder does not continue when there are no MBs
left. Adding a similar check for multi-thread decode as well.

Bug: 69269702
Test: manual
Change-Id: Ibbe5202dbb270625e4f592b4fdb8ef0ec71a979d
(cherry picked from commit 00a2482c8dfa3550bcbfa515a93a4cead5daf8e9)